requirements.txt
alembic
fastapi
httpx
langchain-openai
langchain-core
openai
paypalrestsdk
pydantic
pydantic-settings
python-dotenv
python-multipart
sqlalchemy
sqlmodel
stripe
uvicorn
web3

# Database drivers
asyncpg
psycopg2-binary

# Development dependencies
black
pytest
pytest-asyncio
pytest-cov
pytest-postgresql
ruff
mako
